requirements.txt
# Django
Django==5.0.14
gunicorn==23.0.0

# Database
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# Static files
whitenoise==6.8.2

# Environment variables
python-decouple==3.8

# MLS API
wfrmls==1.3.7

# Filtering
django-filter==24.3

# Utilities
python-dateutil==2.9.0.post0

